AI/ML/requirements.txt
# For serving the API
fastapi
uvicorn[standard]

# For AI/ML
google-generativeai
sentence-transformers
scikit-learn
numpy